+++ graphql/codegen/src/core/codegen/templates/orm-client.ts
@@ -12,13 +12,13 @@ import type {
   GraphQLAdapter,
   GraphQLError,
   QueryResult,
-} from '@constructive-io/graphql-types';
+} from '@constructive-io/graphql-query/runtime';
 
 export type {
   GraphQLAdapter,
   GraphQLError,
   QueryResult,
-} from '@constructive-io/graphql-types';
+} from '@constructive-io/graphql-query/runtime';
 
 /**
  * Default adapter that uses fetch for HTTP requests.
