+++ graphql/codegen/src/core/codegen/templates/query-builder.ts
@@ -8,8 +8,8 @@
  * Any changes here will affect all generated ORM clients.
  */
 
-import { parseType, print } from '@0no-co/graphql.web';
-import * as t from 'gql-ast';
+import { parseType, print } from '@constructive-io/graphql-query/runtime';
+import * as t from '@constructive-io/graphql-query/runtime';
 import type {
   ArgumentNode,
   EnumValueNode,
