+++ graphql/orm-test/__tests__/mega-query.test.ts
@@ -20,7 +20,7 @@ import { join } from 'path';
 import { getConnectionsObject, seed } from 'graphile-test';
 import type { GraphQLQueryFnObj, GraphQLResponse } from 'graphile-test';
 import { ConstructivePreset } from 'graphile-settings';
-import { runCodegenAndLoad } from './helpers/codegen-helper';
+import { runCodegenAndLoad } from '@constructive-io/graphql-test';
 import { GraphileTestAdapter } from './helpers/graphile-adapter';
 
 jest.setTimeout(120000);
