+++ graphql/orm-test/__tests__/postgis-spatial.test.ts
@@ -24,7 +24,7 @@ import { join } from 'path';
 import { getConnectionsObject, seed } from 'graphile-test';
 import type { GraphQLQueryFnObj } from 'graphile-test';
 import { ConstructivePreset } from 'graphile-settings';
-import { runCodegenAndLoad } from './helpers/codegen-helper';
+import { runCodegenAndLoad } from '@constructive-io/graphql-test';
 import { GraphileTestAdapter } from './helpers/graphile-adapter';
 
 jest.setTimeout(120000);
